archiver/requirements.txt
functions-framework==3.5.0
flask==2.3.3
requests==2.31.0
google-cloud-tasks==2.19.2
google-cloud-logging==3.9.0
google-auth==2.29.0
google-cloud-secret-manager
google-cloud-storage==2.14.0
pydantic==2.6.1
python-dotenv==1.0.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0